feat: move keys that are mostly related to faktory to a separate faktory.go file
