+++ pkg/logfields/logfields.go
@@ -16,46 +16,15 @@ func TraceField(traceID string) zapcore.Field {
 
 // these constants represents the keys to get these data fields out of a zap logger.
 const (
-	BatchIDKey string = "BID"
-	JobIDKey   string = "JID"
-	JobTypeKey string = "job_type"
-
-	translatedAuditQueueIDKey = "translated_audit_queue_id"
-	auditQueueAttemptsKey     = "audit_queue_attempts"
-	auditChangeKey            = "change_id"
-	modelPlanIDKey            = "model_plan_id"
-	dateKey                   = "date"
-	userIDKey                 = "user_id"
+	auditChangeKey = "change_id"
+	modelPlanIDKey = "model_plan_id"
+	dateKey        = "date"
+	userIDKey      = "user_id"
 )
-
-// BID returns the zap core field for a worker BatchID
-func BID(bid string) zapcore.Field {
-	return zap.String(BatchIDKey, bid)
-}
-
-// JID returns the zap core field for a worker JobID
-func JID(jid string) zapcore.Field {
-	return zap.String(JobIDKey, jid)
-}
-
-// JobType returns the zap core field for a worker job type
-func JobType(jobType string) zapcore.Field {
-	return zap.String(JobTypeKey, jobType)
-}
-
-// TranslatedAuditQueueID provides the zap field for a TranslatedAuditQueueID
-func TranslatedAuditQueueID(translatedAuditQueueID uuid.UUID) zapcore.Field {
-	return zap.Any(translatedAuditQueueIDKey, translatedAuditQueueID)
-}
 
 // AuditChangeID provides the zap field for an audit change id
 func AuditChangeID(changeID interface{}) zapcore.Field {
 	return zap.Any(auditChangeKey, changeID)
-}
-
-// AuditQueueAttempts provides the zap field for the number of attempts
-func AuditQueueAttempts(attempts interface{}) zapcore.Field {
-	return zap.Any(auditQueueAttemptsKey, attempts)
 }
 
 // Date provides the zap field for the date
